fix: resolve 4 Phase 2 contract blockers
- HealthMetric.swift: add 3 missing HealthCategory enum cases (respiratory_rate, skin_temperature, training_load) per phase-2-swift-models contract
- Goal.swift: change targetValue type from Decimal to Double to match contract
- StravaServiceProtocol: remove async keyword from isConnected property to match contract signature
- strava-sync/index.ts: update StravaSyncResponse fields to match phase-2-strava-metrics contract (activities_synced, metrics_synced instead of message/data)

All fixes verified via xcodebuild — BUILD SUCCEEDED

Verification report added to docs for reference.

+++ HealthOS/Models/Goal.swift
@@ -8,7 +8,7 @@ struct Goal: Identifiable, Codable {
     var title: String
     var description: String?
     var category: GoalCategory
-    var targetValue: Decimal
+    var targetValue: Double
     var targetUnit: String
     var currentValue: Decimal?
     var targetDate: Date?
